requirements_complete.txt
# Elysian Fields - Complete Requirements
# Cemetery Management System Dependencies

# Core Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7

# Image Processing & OCR
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3

# Data Handling
pandas==2.0.3

# Configuration
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# Google Maps Integration
folium==0.20.0
googlemaps==4.10.0

# Desktop GUI
customtkinter==5.2.0

# Production Server (optional)
gunicorn==21.2.0
